create processor
processor is the main processing class. Create the class and basic
functions to implement the remainder in the next steps.

Since the processor runs on a loop that is controlled from the outside,
the tests need to create a thread to stop the execution in the tests,
which requries valgrind to run with a scheduler, otherwise the tests
never end.